feat: Add global display filter state to annotation store
Introduced display filter state and a set display filter action to the annotation store. The details settings panel has been refactored to use this new global state instead of local state for its filter switches. This centralizes the display filter logic, allowing other components in the editor to subscribe and react to changes in visibility for annotations.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>