feat: Implement save functionality by connecting to server action
Implemented the save functionality in the editor header. The handle save function now calculates a changeset by comparing current and original detections from the Zustand store. This changeset is then passed to the save detections server action. On success, the client-side state is synchronized using commit changes function and the TanStack Query cache for the case is invalidated.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>