feat: Add keyboard shortcuts panel to editor sidebar
Introduced a keyboard shortcuts panel to the annotation editor sidebar to improve user discoverability of hotkeys. A details shortcuts panel component was created to display a formatted list of all available shortcuts. The editor sidebar was updated to include a new shortcuts button and the logic to dynamically render this panel. The shortcuts button is responsively hidden on mobile devices.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>