fix: Prevent session monitor from logging out user on network failure
Addressed an issue in the use session monitor hook where a user would be redirected on logout even if the cause was a network failure or unreachable server. Implemented an server reachable check that pings the `/api/auth/session` endpoint before triggering a redirect. The user is now only redirected to the homepage if the session is confirmed to be unauthenticated while the server is responsive, preventing logouts due to temporary network issues.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>