feat: Implement server action for verification status metrics
Added verification status server action to compute detailed verification progress. It aggregates data at three levels. The first is case categorized as verified, unverified, or in progress based on their contained detections. The second is image categorized based on detections within the image. The last one are detections counted as either verified or unverified. This provides a granular view of the user's progress in reviewing analysis results.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>